requirements.txt
# Federated Learning for Disease Outbreak Prediction
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
tensorflow>=2.13.0
tensorflow-federated>=0.62.0

# Federated Learning Frameworks
flwr>=1.5.0
pysyft>=0.7.0

# Privacy & Security
opacus>=1.4.0
cryptography>=41.0.0
phe>=1.5.0

# Data Processing
scipy>=1.11.0
statsmodels>=0.14.0
geopandas>=0.13.0

# API & Web
fastapi>=0.103.0
uvicorn>=0.23.0
streamlit>=1.26.0
dash>=2.12.0
flask>=2.3.0

# Communication
grpcio>=1.57.0
requests>=2.31.0
websockets>=11.0.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
joblib>=1.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Jupyter (for development)
jupyter>=1.0.0
ipywidgets>=8.1.0
